bl/requirements.txt
annotated-types==0.7.0
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
gunicorn==23.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
python-engineio==4.9.1
python-socketio==5.11.3
simple-websocket==1.0.0
typing_extensions==4.12.2
Werkzeug==3.0.4
wsproto==1.2.0
